owfile: Change relocation semantics
If a stored path item names a relative path based on some prefix, then
only resolve based on that prefix (i.e. do not allow absolute path to
override it).

This solves a problem where making a full copy of a directory containing
the workflow file and datasets would (while the original directory
exists) still load the data from the original location.