data.TimeVariable: new type of continuous variable
TimeVariable is a continuous variable with Unix epoch
(1970-01-01 00:00:00+0000) as the origin (0.0). Later dates are
positive real numbers (equivalent to Unix timestamp, with microseconds
in the fraction part), and the dates before it map to the negative
real numbers.

Unfortunately, due to limitation of Python datetime, only dates
with year >= 1 (A.D.) are supported.

Only ISO 8601 date(time) formats are supported.

If time is specified without a date, Unix epoch is assumed.

If time is specified without an UTC offset part, localtime is assumed.