canvas/annotations: Fix jerky arrow annotation editing
The ControlPoint graphics object was relying on the default
QGraphicsItem for movement (QGraphicsItem.ItemIsMovable), but
this moves all items in the scene that are selected and movable
including the selected ArrowAnnotation item. Therefore Qt would move
both items, but the ArrowAnnotation's shape/geometry would then be
'fixed' on the control point's positionChanged signal. However when
processing a mouseMoveEvent which did not actually move from the last
reported position the control point would not signal a changed position
(as it did not change) but Qt's default event handler would still
move/set positions for all selected movable items (relative to the
initial move starting point and not a difference from the last
reported mouse position).

This is fixed by reimplementing ControlPoints's mouseMoveEvent to handle
movement itself.