tests.test_doctest: test our doctests, maintain up-to-date documentation
Includes +ELLIPSES, +NORMALIZE_WHITESPACE, so you can use those.

Contrary to popular use, it makes ALL the doctests in a module
use the SAME execution context (globals), so the tests can reuse
variables/results from the global context, so as to avoid having
to reinstantiate the same stuff all over the place.
(Reuse only the global context; the order of tests execution is
random, so doctests of one can't depend on other unit's doctests.)