LearnerScorer: Change signature of score method
It used to have model as a parameter and now it gets the data directly.
This allows more freedom for individual learners to define how to score
variables.
E.g. Logistic regression and Linear regression models need to normalize
the data first, before the coefficient make sense.
Random forest on the other hand does not require normalization.