Change pylint evaluation function
This function turns the pylint score into a normal issue count, which
can be used for comparing code style on pull requests. Relying only on
code score is not good, since it is possible to introduce new issues as
long as there are enough new lines of code without issues.