owimageviewer: Avoid RuntimeErrors after clearing the scene
Do not use (Qt's) queued connection to set the retrieved image.
When a scene is cleared all the pending tasks are cancelled and target
widgets deleted, however if the signal is already enqueued, when the
scene is cleared, the signal handler would atempt to set the pixmap on
an already deleted widget.

Fixes gh-1316