SelectRows: Index attrs only by visible in set_data
In `set_data` we calculate the index of an attribute that is passed to `add_rows`. Since `add_rows` only consider visible attributes the index has to be calculated only on visible as well.

Otherwise this caused `attr_combo.setCurrentIndex` to be set to values larger than the number of items inside it. Strangely this works but causes `attr_combo.currentText()` to returns an empty string which of course breaks indexing on domain.