TreeViewer: Fix different trees being displayed for same dataset
OWTreeViewer was not deterministic in showing trees. This happened
because the layout was updated w.r.t the graph node edges, which were
stored in a set object, which is unordered.
I assume that a set was used because speed is important, so I've
implemented edges as an ordered dict, to keep the speed benefits as well
as an ordering.