OWMergeData: Refactor merging - clean-up, reimplement in numpy.
Merge data used to call `group_table_indices` that returned all
occurrences of a certain value. Then it put it through a filter
to keep only the first occurrence. This is no longer sensible since
values are now unique.

Function `group_table_indices` is no longer needed by Merge data:
it is now simplified and moved to Venn diagram.

Merging is refactored so that it first computes a left-join. Then
it removes the unneeded rows if we want an inner-join, or adds the
new rows for outer-join. All this is done in numpy. Also, actual
merging of matrices used to go through Python lists, which were
than put back into numpy arrays. This is now all done in numpy.