WebviewWidget: avoid hangs with webengine
Waiting for execution to complete was handled in an unsafe way. As all
js executions set the same flag when they were done, it was possible
that multiple executions set the flag before the thread waiting for them
could have noticed it. As a result, if two or more executions ended at
the same time the first waiting thread to notice that would clear the
flag and remaining threads would keep waiting indefinetly.

Should fix ocasionally stalled PyQt5 builds on travis.