LineEditWFocusOut: Pressing enters selects text
Pressing Enter in line edit should show the user that the data has
been "entered". Selecting the field seems better than clearing the focus
because (1) it allows the field to be easily re-edited (maybe the user
just wanted to see the efects of her change) and because (2) we do not
lose the focus position (moving or field forward or backward is easy).

Selecting text is also QLineEdit's behaviour on keyboard focus.
This is also what QSpinBox does on Enter.