VizRank: Fix race condition in `toggle`
There was a race condition in the `toggle` method where VizRank could
be left without a running worker thread. This could happen between
setting keep_running from False to True and subsequent call to
QThread.start. The worker could read keep_running as False right
before that and not yet exit before the call to start. Then start would
be a noop and the thread would stop right after.

Fix by explicitly joining on the thread on 'pause' as a synchronization
point.