Table.from_table: fix caching with reused ids
In Table.from_table id() of objects are used as cache keys. Because
ids can be reused this can be a problem. If keys were not tuples
we could use WeakKeyDictionary for cache, but here, because they are, I
store weakrefs to original objects as values. The cached value is not
used if the objects these weakrefs are pointing to do not exists anymore.