owcolor: fix handling changed compute_values
Variables stored in context settings were reused when constructing new
domains. Therefore, there were problems if only their compute_values
changed because the old context (matching names and values) still
matched. After this fix the variable descriptors on the input are always
used when creating colored variables instead of the ones stored in the
context.

Also, compute_values do not need to be stored in settings and are
therefore removed.